feat: 🎸 update typescript template
Update, improve, and fix some bugs in TypeScript template
