+++ templates/typescript/.prettierignore
@@ -1,3 +1,8 @@
+# Module
+node_modules/
+
+# Editor settings
+.vscode/settings.json
+
+# Build output
 dist
-.yarn/*
-node_modules
