+++ templates/typescript/.prettierrc
@@ -1,5 +1,7 @@
 {
-	"useTabs": true,
-	"semi": true,
-	"singleQuote": true
+  "semi": true,
+  "singleQuote": true,
+  "bracketSpacing": true,
+  "trailingComma": "es5",
+  "endOfLine": "lf"
 }
