--- templates/typescript/src/Commands/general/ping.ts
@@ -1,13 +1,0 @@
-import { Command } from '../../Interfaces';
-
-export const command: Command = {
-	name: 'ping',
-	description: 'Ping command',
-	aliases: [],
-	usage: 'ping',
-	testOnly: true,
-	permissions: [],
-	run: async (client, message, args) => {
-		message.reply(`${client.ws.ping}`);
-	},
-};
