+++ templates/typescript/src/Interfaces/Config.ts
@@ -1,9 +1,8 @@
 declare namespace NodeJS {
-	export interface ProcessEnv {
-		GITHUB_AUTH_TOKEN: string;
-		TOKEN: string;
-		PREFIX: string;
-		MONGOURI: string;
-		TESTSERVER: string;
-	}
+  export interface ProcessEnv {
+    TOKEN: string;
+    PREFIX: string;
+    MONGOURI: string;
+    TESTSERVER: string;
+  }
 }
