+++ templates/typescript/src/Interfaces/Event.ts
@@ -1,11 +1,11 @@
-import Client from '../Client';
 import { ClientEvents } from 'discord.js';
+import Client from '../Client';
 
 interface Run {
-	(client: Client, ...args: any[]);
+  (client: Client, ...args: any[]);
 }
 
 export interface Event {
-	name: keyof ClientEvents;
-	run: Run;
+  name: keyof ClientEvents;
+  run: Run;
 }
