+++ templates/typescript/src/Interfaces/Slash.ts
@@ -1,15 +1,11 @@
+import { CommandInteraction, ApplicationCommandData } from 'discord.js';
 import Client from '../Client';
-import {
-	CommandInteraction,
-	ApplicationCommandOption,
-	ApplicationCommandData,
-} from 'discord.js';
 
 interface Run {
-	(client: Client, interaction: CommandInteraction, args: string[]);
+  (client: Client, interaction: CommandInteraction, args: string[]);
 }
 
 export type Slash = ApplicationCommandData & {
-	testOnly: boolean;
-	run: Run;
+  testOnly: boolean;
+  run: Run;
 };
