+++ templates/typescript/src/Slash/General/Ping.ts
@@ -1,0 +1,10 @@
+import { Slash } from '../../Interfaces';
+
+export const slash: Slash = {
+  name: 'ping',
+  description: 'Ping command',
+  testOnly: true,
+  run: async (client, interaction, args) => {
+    interaction.followUp({ content: `${client.ws.ping}ms` });
+  },
+};
