--- templates/typescript/tslint.json
@@ -1,3 +1,0 @@
-{
-	"extends": ["tslint:latest", "tslint-config-prettier"]
-}
